+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - Adds client commands: `config`, `config targets`.
 - All actions and outputs are now done against targets in alphabetical
   order.
-- Adds root command `--version` flag to provide build information.
+- Adds osprey `--version` flag to provide build information.
 
 # Release 1.1.0
 - Adds support for `certificate-authority-data` in client config.